LPD-103793 Move the encryption test helpers to object-test-util
